clean up signup UI
