--- frontend-next/utils/maxLengthField.ts
@@ -1,17 +1,0 @@
-export const COMPANY_NAME_LENGTH = 200;
-export const EMAIL_LENGTH = 50;
-export const BIO_LENGTH = 20000;
-export const PHONE_NUMBER_LENGTH = 20;
-export const WEBSITE_LENGTH = 100;
-export const MISSION_LENGTH = 2000;
-export const CULTURE_LENGTH = 2000;
-export const FIRST_NAME_LENGTH = 64;
-export const NEW_PASSWORD_MAX_LENGTH = 16;
-export const LAST_NAME_LENGTH = 64;
-export const CREATE_POST_LENGTH = 100000;
-export const MAX_LENGTH_READ_MORE_BIO = 100;
-export const MAX_LENGTH_READ_MORE_MISSION = 200;
-export const CITY_LENGTH = 150;
-export const ADDRESS_LENGTH = 150;
-export const MESSAGE_MAX_LENGTH = 20000;
-export const SHOW_MSG_SWITCH_MAX_LENGTH = 25;
